Cleaned up scripts
