--- examples/scripts/utils.mjs
@@ -1,27 +1,0 @@
-/**
- * @typedef {object} ExampleConfig
- * @property {string} [DESCRIPTION] - The example description.
- * @property {boolean} [HIDDEN] - The example is hidden on production.
- * @property {'development' | 'performance' | 'debug'} [ENGINE] - The engine type.
- * @property {boolean} [NO_DEVICE_SELECTOR] - No device selector.
- * @property {boolean} [NO_MINISTATS] - No ministats.
- * @property {boolean} [WEBGPU_DISABLED] - If webgpu is disabled.
- * @property {boolean} [WEBGL_DISABLED] - If webgl is disabled.
- */
-
-/**
- * @param {string} script - The script to parse.
- * @returns {ExampleConfig} - The parsed config.
- */
-export const parseConfig = (script) => {
-    const regex = /\/\/ @config (\S+)(?:\s+([^\n]+))?/g;
-    let match;
-    /** @type {Record<string, any>} */
-    const config = {};
-    while ((match = regex.exec(script)) !== null) {
-        const key = match[1].trim();
-        const val = match[2]?.trim();
-        config[key] = /true|false/.test(val) ? val === 'true' : val ?? true;
-    }
-    return config;
-};
